requirements.txt
faiss-cpu
langchain-huggingface
huggingface_hub
pypdf2
langchain-community
streamlit  
asyncio
fitz
fastapi
uvicorn
langchain
openai
faiss-cpu
python-dotenv
openai
sphinx
pyJoules